Update ui and join button
