+++ components/ProposalVotingPower/QuadraticVotingInfoModal.tsx
@@ -35,18 +35,14 @@ export default function QuadraticVotingInfoModal({
           <div className="p-4">
             <div className="bg-bkg-1 p-4 rounded-lg mb-8 flex flex-row justify-evenly">
               <div className="flex flex-col items-end">
-                <p className="font-bold">
-                  {tokenAmount} tokens | {voteWeight} votes
-                </p>
-
-                <p className="text-fgd-3 mb-2">
-                  {(
-                    (Number(voteWeight ?? '0') /
-                      Number(totalVoteWeight ?? '0')) *
-                    100
-                  ).toFixed(2) ?? 0}
-                  % of possible votes
-                </p>
+                <div className="flex-col items-center">
+                  <p className="font-bold mr-2 text-xl">
+                    {voteWeight ?? '0'} votes
+                  </p>
+                  <p className="text-fgd-3 text-xs justify-self">
+                    ({tokenAmount ?? '0'} tokens)
+                  </p>
+                </div>
               </div>
               <UserGroupIcon className="w-12 h-12" />
               <div className="flex flex-col">
